+++ Frameworks/XADMaster.framework/Versions/A/Headers/XADPath.h
@@ -48,7 +48,7 @@ separators:(const char *)separators source:(XADStringSource *)stringsource;
 -(XADStringSource *)source;
 
 -(BOOL)isEqual:(id)other;
--(unsigned)hash;
+-(NSUInteger)hash;
 -(id)copyWithZone:(NSZone *)zone;
 
 @end
